+++ src/keyframes.ts
@@ -1,0 +1,11 @@
+export const fadeOut = {
+  from: { opacity: 1 },
+  to: { opacity: 0 },
+  duration: 350
+}
+
+export const fadeIn = {
+  from: { opacity: 0 },
+  to: { opacity: 1 },
+  duration: 350
+}
